Репозитории и Сервисы юзера и постов

+++ src/main/java/ru/sberbank/homework/dragonblog/service/PostService.java
@@ -7,17 +7,17 @@ import java.time.LocalDateTime;
 import java.util.List;
 
 public interface PostService {
-    Post get(int id) throws NotFoundException;
+    Post get(long id) throws NotFoundException;
 
     // false if post do not belong to author
-    boolean delete(int id, int userId) throws NotFoundException;
+    boolean delete(long id, long userId) throws NotFoundException;
 
     // null if post do not belong to author
-    Post update(Post post, int userId) throws NotFoundException;
+    Post update(Post post, long userId) throws NotFoundException;
 
-    Post create(Post post, int userId);
+    Post create(Post post, long userId);
 
-    List<Post> getAllByUser(int userId);
+    List<Post> getAllByUser(long userId);
 
     //will find in uniquename
     List<Post> getAllBySearchString(String search);
